IS-41:  Modified billable attribute from type int to  type Integer for the purpose of detecting whether or not the client submitted its value.   This is in support of creating message handlers for the GET_TASK transaction code

+++ api_parent/api_projecttracker/src/test/java/org/rmt2/api/admin/ProjectAdminQueryApiTest.java
@@ -15,8 +15,8 @@ import org.dao.mapping.orm.rmt2.VwTimesheetEventList;
 import org.dao.mapping.orm.rmt2.VwTimesheetProjectTask;
 import org.dto.ClientDto;
 import org.dto.EventDto;
-import org.dto.ProjectClientDto;
 import org.dto.Project2Dto;
+import org.dto.ProjectClientDto;
 import org.dto.ProjectEventDto;
 import org.dto.ProjectTaskDto;
 import org.dto.TaskDto;
@@ -464,7 +464,7 @@ public class ProjectAdminQueryApiTest extends ProjectTrackerMockData {
         Assert.assertEquals(1, results.size());
         TaskDto obj = results.get(0);
         Assert.assertEquals(obj.getTaskId(), ProjectTrackerMockDataFactory.TEST_TASK_ID);
-        Assert.assertEquals(obj.getTaskBillable(), 1);
+        Assert.assertEquals(obj.getTaskBillable().intValue(), 1);
         Assert.assertEquals(obj.getTaskDescription(), "Design and Analysis");
     }
